+++ examples/beat_app.rs
@@ -1,7 +1,8 @@
 #![allow(unused_variables)]
 
 use anyhow::Result;
-use celery::beat::RegularSchedule;
+use celery::beat::{CronSchedule, RegularSchedule};
+use celery::broker::AMQPBroker;
 use celery::task::TaskResult;
 use env_logger::Env;
 use tokio::time::Duration;
@@ -20,21 +21,22 @@ fn long_running_task(secs: Option<u64>) -> TaskResult<()> {
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    env_logger::from_env(Env::default().default_filter_or("info")).init();
+    env_logger::Builder::from_env(Env::default().default_filter_or("info")).init();
 
     // Build a `Beat` with a default scheduler backend.
     let mut beat = celery::beat!(
-        broker = AMQP { std::env::var("AMQP_ADDR").unwrap_or_else(|_| "amqp://127.0.0.1:5672/my_vhost".into()) },
+        broker = AMQPBroker { std::env::var("AMQP_ADDR").unwrap_or_else(|_| "amqp://127.0.0.1:5672/my_vhost".into()) },
         task_routes = [
             "*" => QUEUE_NAME,
         ],
-    );
+    ).await?;
 
     // Add scheduled tasks to the default `Beat` and start it.
     let add_schedule = RegularSchedule::new(Duration::from_secs(5));
     beat.schedule_task(add::new(1, 2), add_schedule);
 
-    let long_running_schedule = RegularSchedule::new(Duration::from_secs(10));
+    // The long running task will run every two minutes.
+    let long_running_schedule = CronSchedule::from_string("*/2 * * * *")?;
     beat.schedule_task(long_running_task::new(Some(1)), long_running_schedule);
 
     beat.start().await?;