+++ src/beat/mod.rs
@@ -1,4 +1,4 @@
-//! Celery beat is an app that can automatically produce tasks at scheduled times.
+//! Celery [`Beat`] is an app that can automatically produce tasks at scheduled times.
 //!
 //! ### Terminology
 //!
@@ -39,7 +39,7 @@ mod backend;
 pub use backend::{LocalSchedulerBackend, SchedulerBackend};
 
 mod schedule;
-pub use schedule::{RegularSchedule, Schedule};
+pub use schedule::{CronSchedule, RegularSchedule, Schedule};
 
 mod scheduled_task;
 pub use scheduled_task::ScheduledTask;
@@ -59,7 +59,7 @@ where
     task_options: TaskOptions,
 }
 
-/// Used to create a `Beat` app with a custom configuration.
+/// Used to create a [`Beat`] app with a custom configuration.
 pub struct BeatBuilder<Bb, Sb>
 where
     Bb: BrokerBuilder,
@@ -210,15 +210,16 @@ where
     }
 }
 
-/// A `Beat` app is used to send out scheduled tasks. This is the struct that is
-/// created with the [`beat`](../macro.beat.html) macro.
+/// A [`Beat`] app is used to send out scheduled tasks. This is the struct that is
+/// created with the [`beat!`] macro.
 ///
 /// It drives execution by making the internal scheduler "tick", and updates the list of scheduled
 /// tasks through a customizable scheduler backend.
 pub struct Beat<Br: Broker, Sb: SchedulerBackend> {
     pub name: String,
-    scheduler: Scheduler<Br>,
-    scheduler_backend: Sb,
+    pub scheduler: Scheduler<Br>,
+    pub scheduler_backend: Sb,
+
     task_routes: Vec<Rule>,
     default_queue: String,
     task_options: TaskOptions,
@@ -314,7 +315,7 @@ where
             let mut reconnect_successful: bool = false;
             for _ in 0..self.broker_connection_max_retries {
                 info!("Trying to re-establish connection with broker");
-                time::delay_for(Duration::from_secs(
+                time::sleep(Duration::from_secs(
                     self.broker_connection_retry_delay as u64,
                 ))
                 .await;
@@ -360,7 +361,7 @@ where
                     "Unexpected error when unwrapping a SystemTime comparison that is not supposed to fail",
                 );
                 debug!("Now sleeping for {:?}", sleep_interval);
-                time::delay_for(sleep_interval).await;
+                time::sleep(sleep_interval).await;
             }
         }
     }