+++ src/export.rs
@@ -1,4 +1,5 @@
 pub use crate::async_trait::async_trait;
-pub use crate::futures::executor::block_on;
-pub use crate::once_cell::sync::OnceCell;
 pub use crate::serde::{Deserialize, Serialize};
+pub use crate::tokio::runtime::Runtime;
+pub use std::sync::Arc;
+pub type Result<T> = std::result::Result<T, crate::error::CeleryError>;