+++ src/lib.rs
@@ -3,7 +3,7 @@
 //!
 //! # Examples
 //!
-//! Define tasks by decorating functions with the [`task`](attr.task.html) attribute:
+//! Define tasks by decorating functions with the [`macro@task`] attribute:
 //!
 //! ```rust
 //! use celery::prelude::*;
@@ -14,56 +14,66 @@
 //! }
 //! ```
 //!
-//! Then create a [`Celery`](struct.Celery.html) app with the [`app`](macro.app.html)
+//! Then create a [`Celery`] app with the [`app!`]
 //! macro and register your tasks with it:
 //!
 //! ```rust,no_run
+//! # use anyhow::Result;
+//! # use celery::prelude::*;
 //! # #[celery::task]
 //! # fn add(x: i32, y: i32) -> celery::task::TaskResult<i32> {
 //! #     Ok(x + y)
 //! # }
+//! # #[tokio::main]
+//! # async fn main() -> Result<()> {
 //! let my_app = celery::app!(
-//!     broker = AMQP { std::env::var("AMQP_ADDR").unwrap() },
+//!     broker = AMQPBroker { std::env::var("AMQP_ADDR").unwrap() },
 //!     tasks = [add],
 //!     task_routes = [],
-//! );
+//! ).await?;
+//! # Ok(())
+//! # }
 //! ```
 //!
-//! The Celery app can be used as either a producer or consumer (worker). To send tasks to a
-//! queue for a worker to consume, use the [`Celery::send_task`](struct.Celery.html#method.send_task) method:
+//! The [`Celery`] app can be used as either a producer or consumer (worker). To send tasks to a
+//! queue for a worker to consume, use the [`Celery::send_task`] method:
 //!
 //! ```rust,no_run
+//! # use anyhow::Result;
+//! # use celery::prelude::*;
 //! # #[celery::task]
 //! # fn add(x: i32, y: i32) -> celery::task::TaskResult<i32> {
 //! #     Ok(x + y)
 //! # }
 //! # #[tokio::main]
-//! # async fn main() -> anyhow::Result<()> {
+//! # async fn main() -> Result<()> {
 //! # let my_app = celery::app!(
-//! #     broker = AMQP { std::env::var("AMQP_ADDR").unwrap() },
+//! #     broker = AMQPBroker { std::env::var("AMQP_ADDR").unwrap() },
 //! #     tasks = [add],
 //! #     task_routes = [],
-//! # );
+//! # ).await?;
 //! my_app.send_task(add::new(1, 2)).await?;
-//! #   Ok(())
+//! # Ok(())
 //! # }
 //! ```
 //!
-//! And to act as worker and consume tasks sent to a queue by a producer, use the
-//! [`Celery::consume`](struct.Celery.html#method.consume) method:
+//! And to act as a worker to consume tasks sent to a queue by a producer, use the
+//! [`Celery::consume`] method:
 //!
 //! ```rust,no_run
+//! # use anyhow::Result;
+//! # use celery::prelude::*;
 //! # #[celery::task]
 //! # fn add(x: i32, y: i32) -> celery::task::TaskResult<i32> {
 //! #     Ok(x + y)
 //! # }
 //! # #[tokio::main]
-//! # async fn main() -> anyhow::Result<()> {
+//! # async fn main() -> Result<()> {
 //! # let my_app = celery::app!(
-//! #     broker = AMQP { std::env::var("AMQP_ADDR").unwrap() },
+//! #     broker = AMQPBroker { std::env::var("AMQP_ADDR").unwrap() },
 //! #     tasks = [add],
 //! #     task_routes = [],
-//! # );
+//! # ).await?;
 //! my_app.consume().await?;
 //! # Ok(())
 //! # }
@@ -89,7 +99,7 @@ pub mod task;
 #[cfg(feature = "codegen")]
 mod codegen;
 
-/// A procedural macro for generating a [`Task`](task/trait.Task.html) from a function.
+/// A procedural macro for generating a [`Task`](crate::task::Task) from a function.
 ///
 /// If the annotated function has a return value, the return value must be a
 /// [`TaskResult<R>`](task/type.TaskResult.html).
@@ -110,7 +120,7 @@ mod codegen;
 /// first argument should be a reference to `Self`. Note however that Rust won't allow you to call
 /// the argument `self`. Instead, you could use `task` or just `t`.
 /// - `on_failure`: An async callback function to run when the task fails. Should accept a reference to
-/// a task instance and a reference to a `TaskError`.
+/// a task instance and a reference to a [`TaskError`](error/enum.TaskError.html).
 /// - `on_success`: An async callback function to run when the task succeeds. Should accept a reference to
 /// a task instance and a reference to the value returned by the task.
 ///
@@ -190,13 +200,10 @@ pub use codegen::task;
 pub mod export;
 
 #[cfg(feature = "codegen")]
-extern crate futures;
-
-#[cfg(feature = "codegen")]
-extern crate once_cell;
-
-#[cfg(feature = "codegen")]
 extern crate async_trait;
 
 #[cfg(feature = "codegen")]
 extern crate serde;
+
+#[cfg(feature = "codegen")]
+extern crate tokio;